requirements.txt
fastapi
uvicorn[standard]
pydantic
sse-starlette
openai
python-dotenv
requests
pymongo
python-jose[cryptography]
passlib[bcrypt]
